Moved IntoLuaAst trait and impls into own file
