+++ c2rust-refactor/src/scripting.rs
@@ -14,7 +14,7 @@ use rlua::prelude::{LuaContext, LuaError, LuaFunction, LuaResult, LuaTable};
 use rlua::{Lua, UserData, UserDataMethods};
 use rustc_interface::interface;
 use slotmap::{new_key_type, SlotMap};
-use syntax::ast::{self, ExprKind};
+use syntax::ast;
 use syntax::ptr::P;
 
 use crate::command::{self, CommandState, RefactorState};
@@ -24,10 +24,11 @@ use crate::matcher::{self, mut_visit_match_with, Bindings, MatchCtxt, Pattern, S
 use crate::RefactorCtxt;
 
 pub mod ast_visitor;
+pub mod into_lua_ast;
 pub mod utils;
 
 use ast_visitor::LuaAstVisitor;
-use utils::iter_to_lua_array;
+use into_lua_ast::IntoLuaAst;
 
 /// Refactoring module
 // @module Refactor
@@ -60,353 +61,6 @@ pub fn run_lua_file(
     .unwrap_or_else(|e| panic!("User script failed: {:#?}", e));
 
     Ok(())
-}
-
-trait IntoLuaAst<'lua> {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>>;
-}
-
-/// AST Stmt
-// @table Stmt
-// @field type "Stmt"
-// @tfield string kind `StmtKind` of this statement
-//
-// `StmtKind::Local` only:
-// @tfield[opt] LuaAstNode ty Type of local
-// @tfield[opt] LuaAstNode init Initializer of local
-// @tfield LuaAstNode pat Name of local
-//
-// `StmtKind::Item` only:
-// @tfield LuaAstNode item Item node
-//
-// `StmtKind::Semi` and `StmtKind::Expr` only:
-// @tfield LuaAstNode expr Expression in this statement
-impl<'lua> IntoLuaAst<'lua> for ast::Stmt {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-        ast.set("type", "Stmt")?;
-        match self.node {
-            ast::StmtKind::Local(l) => {
-                ast.set("kind", "Local")?;
-                let ast::Local { pat, ty, init, .. } = l.into_inner();
-
-                ast.set("pat", pat.into_lua_ast(ctx, lua_ctx)?)?;
-
-                if let Some(ty) = ty {
-                    ast.set("ty", ctx.intern(ty))?;
-                }
-
-                if let Some(init) = init {
-                    ast.set("init", init.into_lua_ast(ctx, lua_ctx)?)?;
-                }
-            }
-            ast::StmtKind::Item(i) => {
-                ast.set("kind", "Item")?;
-                ast.set("item", i.into_lua_ast(ctx, lua_ctx)?)?;
-            }
-            ast::StmtKind::Semi(e) => {
-                ast.set("kind", "Semi")?;
-                ast.set("expr_old", ctx.intern(e.clone()))?;
-                ast.set("expr", e.into_lua_ast(ctx, lua_ctx)?)?;
-            }
-            ast::StmtKind::Expr(e) => {
-                ast.set("kind", "Expr")?;
-                ast.set("expr_old", ctx.intern(e.clone()))?;
-                ast.set("expr", e.into_lua_ast(ctx, lua_ctx)?)?;
-            }
-            ast::StmtKind::Mac(_) => {
-                return Err(LuaError::external("StmtKind::Mac is not yet implemented"));
-            }
-        }
-
-        Ok(ast)
-    }
-}
-
-/// AST Expr
-// @table Expr
-// @field type "Expr"
-// @tfield string kind `ExprKind` of this expression
-//
-// `ExprKind::Lit` only:
-// @field value Literal value of this expression
-impl<'lua> IntoLuaAst<'lua> for P<ast::Expr> {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-        ast.set("type", "Expr")?;
-        ast.set("id", self.id.as_u32())?;
-
-        self.and_then(|expr| {
-            match expr.node {
-                ExprKind::Lit(l) => {
-                    ast.set("kind", "Lit")?;
-                    match l.node {
-                        ast::LitKind::Str(s, _) => ast.set("value", s.to_string())?,
-                        ast::LitKind::Int(i, _) => ast.set("value", i)?,
-                        _ => {
-                            return Err(LuaError::external(format!(
-                                "{:?} is not yet implemented",
-                                l.node
-                            )));
-                        }
-                    }
-                },
-                ExprKind::Box(boxed) => {
-                    ast.set("kind", "Box")?;
-                    ast.set("boxed", boxed.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::ObsoleteInPlace(_, _) => {
-                    ast.set("kind", "ObsoleteInPlace")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Array(values) => {
-                    let vals: LuaResult<Vec<_>> = values
-                        .into_iter()
-                        .map(|v| v.into_lua_ast(ctx, lua_ctx))
-                        .collect();
-
-                    ast.set("kind", "Array")?;
-                    ast.set("values", iter_to_lua_array(vals?.into_iter(), lua_ctx)?)?;
-                },
-                ExprKind::Call(path, args) => {
-                   let args: LuaResult<Vec<_>> = args
-                        .into_iter()
-                        .map(|v| v.into_lua_ast(ctx, lua_ctx))
-                        .collect();
-
-                    ast.set("kind", "Call")?;
-                    ast.set("path", path.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("args", iter_to_lua_array(args?.into_iter(), lua_ctx)?)?;
-                },
-                ExprKind::MethodCall(_, _) => {
-                    ast.set("kind", "MethodCall")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Tup(_) => {
-                    ast.set("kind", "Tup")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Binary(op, lhs, rhs) => {
-                    ast.set("kind", "Binary")?;
-                    ast.set("op", op.node.to_string())?;
-                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Unary(op, expr) => {
-                    ast.set("kind", "Unary")?;
-                    ast.set("op", syntax::ast::UnOp::to_string(op))?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Cast(_, _) => {
-                    ast.set("kind", "Cast")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Type(expr, _ty) => {
-                    ast.set("kind", "Type")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::If(cond, then, opt_else) => {
-                    ast.set("kind", "If")?;
-                    ast.set("cond", cond.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("then", then.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(els) = opt_else {
-                        ast.set("else", els.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-                },
-                ExprKind::IfLet(_, expr, then, opt_else) => {
-                    ast.set("kind", "IfLet")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("then", then.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(els) = opt_else {
-                        ast.set("else", els.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-
-                    // TODO: Flesh out further
-                },
-                ExprKind::While(cond, _, _) => {
-                    ast.set("kind", "While")?;
-                    ast.set("cond", cond.into_lua_ast(ctx, lua_ctx)?)?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::WhileLet(_pats, expr, block, opt_label) => {
-                    ast.set("kind", "WhileLet")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::ForLoop(_pat, expr, block, opt_label) => {
-                    ast.set("kind", "ForLoop")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::Loop(block, opt_label) => {
-                    ast.set("kind", "Loop")?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::Match(expr, _arms) => {
-                    ast.set("kind", "Match")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Closure(capture_by, _is_async, movability, fn_decl, expr, _span) => {
-                    ast.set("kind", "Closure")?;
-                    ast.set("fn_decl", fn_decl.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("capture_by", match capture_by {
-                        ast::CaptureBy::Ref => "Ref",
-                        ast::CaptureBy::Value => "Value",
-                    })?;
-                    ast.set("movability", match movability {
-                        ast::Movability::Movable => "Movable",
-                        ast::Movability::Static => "Static",
-                    })?;
-                },
-                ExprKind::Block(block, opt_label) => {
-                    ast.set("kind", "Block")?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::Async(..) => {
-                    ast.set("kind", "Async")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::TryBlock(block) => {
-                    ast.set("kind", "TryBlock")?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Assign(lhs, rhs) => {
-                    ast.set("kind", "Assign")?;
-                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::AssignOp(op, lhs, rhs) => {
-                    ast.set("kind", "AssignOp")?;
-                    ast.set("op", op.node.to_string())?;
-                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Field(..) => {
-                    ast.set("kind", "Field")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Index(indexed, index) => {
-                    ast.set("kind", "Index")?;
-                    ast.set("indexed", indexed.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("index", index.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Range(opt_lhs, opt_rhs, limits) => {
-                    ast.set("kind", "Range")?;
-
-                    if let Some(lhs) = opt_lhs {
-                        ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-
-                    if let Some(rhs) = opt_rhs {
-                        ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-
-                    ast.set("limits", match limits {
-                        ast::RangeLimits::HalfOpen => "HalfOpen",
-                        ast::RangeLimits::Closed => "Closed",
-                    })?;
-                },
-                ExprKind::Path(_, path) => {
-                    let segments = path
-                        .segments
-                        .into_iter()
-                        .map(|s| s.ident.to_string());
-
-                    ast.set("kind", "Path")?;
-                    ast.set("segments", iter_to_lua_array(segments, lua_ctx)?)?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::AddrOf(mutability, expr) => {
-                    ast.set("kind", "AddrOf")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("mutability", match mutability {
-                        ast::Mutability::Immutable => "Immutable",
-                        ast::Mutability::Mutable => "Mutable",
-                    })?;
-                },
-                ExprKind::Break(opt_label, opt_expr) => {
-                    ast.set("kind", "Break")?;
-
-                    if let Some(expr) = opt_expr {
-                        ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::Continue(opt_label) => {
-                    ast.set("kind", "Continue")?;
-
-                    if let Some(label) = opt_label {
-                        ast.set("label", label.ident.name.as_str().get())?;
-                    }
-                },
-                ExprKind::Ret(opt_val) => {
-                    ast.set("kind", "Ret")?;
-
-                    if let Some(val) = opt_val {
-                        ast.set("value", val.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-                },
-                ExprKind::InlineAsm(..) => {
-                    ast.set("kind", "InlineAsm")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Mac(..) => {
-                    ast.set("kind", "Mac")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Struct(..) => {
-                    ast.set("kind", "Struct")?;
-                    // TODO: Flesh out further
-                },
-                ExprKind::Repeat(expr, _anon_const) => {
-                    ast.set("kind", "Repeat")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Paren(expr) => {
-                    ast.set("kind", "Paren")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Try(expr) => {
-                    ast.set("kind", "Try")?;
-                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ExprKind::Yield(opt_val) => {
-                    ast.set("kind", "Yield")?;
-
-                    if let Some(val) = opt_val {
-                        ast.set("value", val.into_lua_ast(ctx, lua_ctx)?)?;
-                    }
-                },
-                ExprKind::Err => {
-                    ast.set("kind", "Err")?;
-                },
-            }
-
-            Ok(ast)
-        })
-    }
 }
 
 /// Refactoring context
