+++ c2rust-refactor/src/scripting/ast_visitor.rs
@@ -1,14 +1,14 @@
-use rlua::prelude::{LuaContext, LuaError, LuaFunction, LuaResult, LuaTable};
+use rlua::prelude::{LuaContext, LuaFunction, LuaResult, LuaTable};
 use syntax::ast::{
-    Arg, BindingMode, Block, Crate, Expr, ExprKind, FunctionRetTy, FnDecl, ImplItem,
-    ImplItemKind, Item, ItemKind, Local, Mod, Mutability::*, NodeId, Pat, PatKind,
+    Arg, BindingMode, Block, Crate, Expr, ExprKind, FnDecl, ImplItem, ImplItemKind,
+    Item, ItemKind, Local, Mod, Mutability::*, NodeId, Pat, PatKind,
     Stmt, StmtKind,
 };
 use syntax::source_map::symbol::Symbol;
 use syntax::ptr::P;
 
 use crate::ast_manip::fn_edit::{FnKind, FnLike};
-use crate::scripting::{IntoLuaAst, TransformCtxt, utils::iter_to_lua_array};
+use crate::scripting::{TransformCtxt, into_lua_ast::IntoLuaAst};
 
 macro_rules! call_lua_visitor_method {
     ($obj: expr , $method: ident ($($params: expr),*)) => {
@@ -195,222 +195,6 @@ impl<'a, 'lua, 'tctx> LuaAstVisitor<'a, 'lua, 'tctx> {
         call_lua_visitor_method!(self.visitor,finish());
 
         Ok(())
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for FnLike {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "FnLike")?;
-        ast.set("kind", match self.kind {
-            FnKind::Normal => "Normal",
-            FnKind::ImplMethod => "ImplMethod",
-            FnKind::TraitMethod => "TraitMethod",
-            FnKind::Foreign => "Foreign",
-        })?;
-        ast.set("id", self.id.as_u32())?;
-        ast.set("ident", self.ident.as_str().get())?;
-        ast.set("decl", self.decl.into_lua_ast(ctx, lua_ctx)?)?;
-
-        let block = self.block
-            .map(|b| b.into_lua_ast(ctx, lua_ctx))
-            .transpose()?;
-
-        ast.set("block", block)?;
-
-        Ok(ast)
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for P<FnDecl> {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "FnDecl")?;
-        ast.set("c_variadic", self.c_variadic)?;
-
-        self.and_then(|fn_decl| {
-            ast.set("return_type", match fn_decl.output {
-                FunctionRetTy::Default(_) => None,
-                FunctionRetTy::Ty(ty) => Some(ctx.intern(ty)),
-            })?;
-
-            let args: LuaResult<Vec<_>> = fn_decl.inputs
-                .into_iter()
-                .map(|arg| arg.into_lua_ast(ctx, lua_ctx))
-                .collect();
-
-            ast.set("args", iter_to_lua_array(args?.into_iter(), lua_ctx)?)?;
-
-            // TODO: self, self kind
-
-            Ok(ast)
-        })
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for Arg {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "Arg")?;
-        ast.set("id", self.id.as_u32())?;
-        ast.set("pat", self.pat.into_lua_ast(ctx, lua_ctx)?)?;
-
-        Ok(ast)
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for P<Block> {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "Block")?;
-
-        self.and_then(|block| {
-            let stmts = block.stmts
-                .into_iter()
-                .map(|stmt| stmt.into_lua_ast(ctx, lua_ctx))
-                .collect::<LuaResult<Vec<_>>>();
-
-            ast.set("stmts", iter_to_lua_array(stmts?.into_iter(), lua_ctx)?)?;
-
-            Ok(ast)
-        })
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for P<Pat> {
-    fn into_lua_ast(self, _ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "Pat")?;
-        ast.set("id", self.id.as_u32())?;
-
-        self.and_then(|pat| {
-            match pat.node {
-                PatKind::Wild => {
-                    ast.set("kind", "Wild")?;
-                },
-                PatKind::Ident(binding, ident, _sub_pattern) => {
-                    ast.set("kind", "Ident")?;
-                    ast.set("binding", match binding {
-                        BindingMode::ByRef(Immutable) => "ByRefImmutable",
-                        BindingMode::ByRef(Mutable) => "ByRefMutable",
-                        BindingMode::ByValue(Immutable) => "ByValueImmutable",
-                        BindingMode::ByValue(Mutable) => "ByValueMutable",
-                    })?;
-                    ast.set("ident", ident.as_str().get())?;
-                },
-                _ => return Err(LuaError::external(format!("unimplemented pattern type: {:?}", pat.node))),
-            }
-
-            Ok(ast)
-        })
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for Crate {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "Crate")?;
-        ast.set("module", self.module.into_lua_ast(ctx, lua_ctx)?)?;
-
-        Ok(ast)
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for Mod {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-        let items = self.items
-            .into_iter()
-            .map(|item| item.into_lua_ast(ctx, lua_ctx))
-            .collect::<LuaResult<Vec<_>>>();
-
-        ast.set("type", "Mod")?;
-        ast.set("inline", self.inline)?;
-        ast.set("items", iter_to_lua_array(items?.into_iter(), lua_ctx)?)?;
-
-        Ok(ast)
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for P<Item> {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "Item")?;
-
-        self.and_then(|item| {
-            ast.set("id", item.id.as_u32())?;
-            ast.set("ident", item.ident.name.as_str().get())?;
-
-            match item.node {
-                ItemKind::ExternCrate(opt_name) => {
-                    ast.set("kind", "ExternCrate")?;
-
-                    if let Some(name) = opt_name {
-                        ast.set("name", name.as_str().get())?;
-                    }
-                },
-                ItemKind::Mod(module) => {
-                    ast.set("kind", "Mod")?;
-                    ast.set("mod", module.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ItemKind::Use(_use_tree) => {
-                    ast.set("kind", "Use")?;
-                    // ast.set("use_tree", use_tree.into_lua_ast(ctx, lua_ctx))?;
-                },
-                ItemKind::Fn(decl, _header, _generics, block) => {
-                    ast.set("kind", "Fn")?;
-                    ast.set("decl", decl.into_lua_ast(ctx, lua_ctx)?)?;
-                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-                },
-                ItemKind::Struct(_variant_data, _generics) => {
-                    ast.set("kind", "Struct")?;
-                    // TODO: Variant data, generics
-                },
-                ItemKind::Impl(.., items) => {
-                    let items = items
-                        .into_iter()
-                        .map(|item| item.into_lua_ast(ctx, lua_ctx))
-                        .collect::<LuaResult<Vec<_>>>();
-
-                    ast.set("kind", "Impl")?;
-                    ast.set("items", iter_to_lua_array(items?.into_iter(), lua_ctx)?)?;
-
-                    // TODO: other fields
-                },
-                ref e => warn!("Item IntoLuaAst kind: {:?}", e),
-            }
-
-            Ok(ast)
-        })
-
-    }
-}
-
-impl<'lua> IntoLuaAst<'lua> for ImplItem {
-    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
-        let ast = lua_ctx.create_table()?;
-
-        ast.set("type", "ImplItem")?;
-
-        match self.node {
-            ImplItemKind::Method(sig, block) => {
-                ast.set("kind", "ImplMethod")?;
-                ast.set("ident", self.ident.to_string())?;
-                ast.set("decl", sig.decl.into_lua_ast(ctx, lua_ctx)?)?;
-                ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
-                // TODO: generics, attrs, ..
-            },
-            ref e => unimplemented!("IntoLuaAst for {:?}", e),
-        }
-
-        Ok(ast)
     }
 }
