+++ c2rust-refactor/src/scripting/into_lua_ast.rs
@@ -1,0 +1,573 @@
+use rlua::prelude::{LuaContext, LuaError, LuaResult, LuaTable};
+use syntax::ast::{
+    Arg, BindingMode, Block, CaptureBy, Crate, Expr, ExprKind, FunctionRetTy, FnDecl,
+    ImplItem, ImplItemKind, Item, ItemKind, LitKind, Local, Mod, Movability,
+    Mutability::*, Pat, PatKind, RangeLimits, Stmt, StmtKind,
+};
+use syntax::ptr::P;
+
+use crate::ast_manip::fn_edit::{FnKind, FnLike};
+use crate::scripting::{TransformCtxt, utils::iter_to_lua_array};
+
+pub(crate) trait IntoLuaAst<'lua> {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>>;
+}
+
+/// AST Stmt
+// @table Stmt
+// @field type "Stmt"
+// @tfield string kind `StmtKind` of this statement
+//
+// `StmtKind::Local` only:
+// @tfield[opt] LuaAstNode ty Type of local
+// @tfield[opt] LuaAstNode init Initializer of local
+// @tfield LuaAstNode pat Name of local
+//
+// `StmtKind::Item` only:
+// @tfield LuaAstNode item Item node
+//
+// `StmtKind::Semi` and `StmtKind::Expr` only:
+// @tfield LuaAstNode expr Expression in this statement
+impl<'lua> IntoLuaAst<'lua> for Stmt {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+        ast.set("type", "Stmt")?;
+        match self.node {
+            StmtKind::Local(l) => {
+                ast.set("kind", "Local")?;
+                let Local { pat, ty, init, .. } = l.into_inner();
+
+                ast.set("pat", pat.into_lua_ast(ctx, lua_ctx)?)?;
+
+                if let Some(ty) = ty {
+                    ast.set("ty", ctx.intern(ty))?;
+                }
+
+                if let Some(init) = init {
+                    ast.set("init", init.into_lua_ast(ctx, lua_ctx)?)?;
+                }
+            }
+            StmtKind::Item(i) => {
+                ast.set("kind", "Item")?;
+                ast.set("item", i.into_lua_ast(ctx, lua_ctx)?)?;
+            }
+            StmtKind::Semi(e) => {
+                ast.set("kind", "Semi")?;
+                ast.set("expr_old", ctx.intern(e.clone()))?;
+                ast.set("expr", e.into_lua_ast(ctx, lua_ctx)?)?;
+            }
+            StmtKind::Expr(e) => {
+                ast.set("kind", "Expr")?;
+                ast.set("expr_old", ctx.intern(e.clone()))?;
+                ast.set("expr", e.into_lua_ast(ctx, lua_ctx)?)?;
+            }
+            StmtKind::Mac(_) => {
+                return Err(LuaError::external("StmtKind::Mac is not yet implemented"));
+            }
+        }
+
+        Ok(ast)
+    }
+}
+
+/// AST Expr
+// @table Expr
+// @field type "Expr"
+// @tfield string kind `ExprKind` of this expression
+//
+// `ExprKind::Lit` only:
+// @field value Literal value of this expression
+impl<'lua> IntoLuaAst<'lua> for P<Expr> {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+        ast.set("type", "Expr")?;
+        ast.set("id", self.id.as_u32())?;
+
+        self.and_then(|expr| {
+            match expr.node {
+                ExprKind::Lit(l) => {
+                    ast.set("kind", "Lit")?;
+                    match l.node {
+                        LitKind::Str(s, _) => ast.set("value", s.to_string())?,
+                        LitKind::Int(i, _) => ast.set("value", i)?,
+                        _ => {
+                            return Err(LuaError::external(format!(
+                                "{:?} is not yet implemented",
+                                l.node
+                            )));
+                        }
+                    }
+                },
+                ExprKind::Box(boxed) => {
+                    ast.set("kind", "Box")?;
+                    ast.set("boxed", boxed.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::ObsoleteInPlace(_, _) => {
+                    ast.set("kind", "ObsoleteInPlace")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Array(values) => {
+                    let vals: LuaResult<Vec<_>> = values
+                        .into_iter()
+                        .map(|v| v.into_lua_ast(ctx, lua_ctx))
+                        .collect();
+
+                    ast.set("kind", "Array")?;
+                    ast.set("values", iter_to_lua_array(vals?.into_iter(), lua_ctx)?)?;
+                },
+                ExprKind::Call(path, args) => {
+                   let args: LuaResult<Vec<_>> = args
+                        .into_iter()
+                        .map(|v| v.into_lua_ast(ctx, lua_ctx))
+                        .collect();
+
+                    ast.set("kind", "Call")?;
+                    ast.set("path", path.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("args", iter_to_lua_array(args?.into_iter(), lua_ctx)?)?;
+                },
+                ExprKind::MethodCall(_, _) => {
+                    ast.set("kind", "MethodCall")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Tup(_) => {
+                    ast.set("kind", "Tup")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Binary(op, lhs, rhs) => {
+                    ast.set("kind", "Binary")?;
+                    ast.set("op", op.node.to_string())?;
+                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Unary(op, expr) => {
+                    ast.set("kind", "Unary")?;
+                    ast.set("op", syntax::ast::UnOp::to_string(op))?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Cast(_, _) => {
+                    ast.set("kind", "Cast")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Type(expr, _ty) => {
+                    ast.set("kind", "Type")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::If(cond, then, opt_else) => {
+                    ast.set("kind", "If")?;
+                    ast.set("cond", cond.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("then", then.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(els) = opt_else {
+                        ast.set("else", els.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+                },
+                ExprKind::IfLet(_, expr, then, opt_else) => {
+                    ast.set("kind", "IfLet")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("then", then.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(els) = opt_else {
+                        ast.set("else", els.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+
+                    // TODO: Flesh out further
+                },
+                ExprKind::While(cond, _, _) => {
+                    ast.set("kind", "While")?;
+                    ast.set("cond", cond.into_lua_ast(ctx, lua_ctx)?)?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::WhileLet(_pats, expr, block, opt_label) => {
+                    ast.set("kind", "WhileLet")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::ForLoop(_pat, expr, block, opt_label) => {
+                    ast.set("kind", "ForLoop")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::Loop(block, opt_label) => {
+                    ast.set("kind", "Loop")?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::Match(expr, _arms) => {
+                    ast.set("kind", "Match")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Closure(capture_by, _is_async, movability, fn_decl, expr, _span) => {
+                    ast.set("kind", "Closure")?;
+                    ast.set("fn_decl", fn_decl.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("capture_by", match capture_by {
+                        CaptureBy::Ref => "Ref",
+                        CaptureBy::Value => "Value",
+                    })?;
+                    ast.set("movability", match movability {
+                        Movability::Movable => "Movable",
+                        Movability::Static => "Static",
+                    })?;
+                },
+                ExprKind::Block(block, opt_label) => {
+                    ast.set("kind", "Block")?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::Async(..) => {
+                    ast.set("kind", "Async")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::TryBlock(block) => {
+                    ast.set("kind", "TryBlock")?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Assign(lhs, rhs) => {
+                    ast.set("kind", "Assign")?;
+                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::AssignOp(op, lhs, rhs) => {
+                    ast.set("kind", "AssignOp")?;
+                    ast.set("op", op.node.to_string())?;
+                    ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Field(..) => {
+                    ast.set("kind", "Field")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Index(indexed, index) => {
+                    ast.set("kind", "Index")?;
+                    ast.set("indexed", indexed.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("index", index.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Range(opt_lhs, opt_rhs, limits) => {
+                    ast.set("kind", "Range")?;
+
+                    if let Some(lhs) = opt_lhs {
+                        ast.set("lhs", lhs.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+
+                    if let Some(rhs) = opt_rhs {
+                        ast.set("rhs", rhs.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+
+                    ast.set("limits", match limits {
+                        RangeLimits::HalfOpen => "HalfOpen",
+                        RangeLimits::Closed => "Closed",
+                    })?;
+                },
+                ExprKind::Path(_, path) => {
+                    let segments = path
+                        .segments
+                        .into_iter()
+                        .map(|s| s.ident.to_string());
+
+                    ast.set("kind", "Path")?;
+                    ast.set("segments", iter_to_lua_array(segments, lua_ctx)?)?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::AddrOf(mutability, expr) => {
+                    ast.set("kind", "AddrOf")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("mutability", match mutability {
+                        Immutable => "Immutable",
+                        Mutable => "Mutable",
+                    })?;
+                },
+                ExprKind::Break(opt_label, opt_expr) => {
+                    ast.set("kind", "Break")?;
+
+                    if let Some(expr) = opt_expr {
+                        ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::Continue(opt_label) => {
+                    ast.set("kind", "Continue")?;
+
+                    if let Some(label) = opt_label {
+                        ast.set("label", label.ident.name.as_str().get())?;
+                    }
+                },
+                ExprKind::Ret(opt_val) => {
+                    ast.set("kind", "Ret")?;
+
+                    if let Some(val) = opt_val {
+                        ast.set("value", val.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+                },
+                ExprKind::InlineAsm(..) => {
+                    ast.set("kind", "InlineAsm")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Mac(..) => {
+                    ast.set("kind", "Mac")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Struct(..) => {
+                    ast.set("kind", "Struct")?;
+                    // TODO: Flesh out further
+                },
+                ExprKind::Repeat(expr, _anon_const) => {
+                    ast.set("kind", "Repeat")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Paren(expr) => {
+                    ast.set("kind", "Paren")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Try(expr) => {
+                    ast.set("kind", "Try")?;
+                    ast.set("expr", expr.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ExprKind::Yield(opt_val) => {
+                    ast.set("kind", "Yield")?;
+
+                    if let Some(val) = opt_val {
+                        ast.set("value", val.into_lua_ast(ctx, lua_ctx)?)?;
+                    }
+                },
+                ExprKind::Err => {
+                    ast.set("kind", "Err")?;
+                },
+            }
+
+            Ok(ast)
+        })
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for FnLike {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "FnLike")?;
+        ast.set("kind", match self.kind {
+            FnKind::Normal => "Normal",
+            FnKind::ImplMethod => "ImplMethod",
+            FnKind::TraitMethod => "TraitMethod",
+            FnKind::Foreign => "Foreign",
+        })?;
+        ast.set("id", self.id.as_u32())?;
+        ast.set("ident", self.ident.as_str().get())?;
+        ast.set("decl", self.decl.into_lua_ast(ctx, lua_ctx)?)?;
+
+        let block = self.block
+            .map(|b| b.into_lua_ast(ctx, lua_ctx))
+            .transpose()?;
+
+        ast.set("block", block)?;
+
+        Ok(ast)
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for P<FnDecl> {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "FnDecl")?;
+        ast.set("c_variadic", self.c_variadic)?;
+
+        self.and_then(|fn_decl| {
+            ast.set("return_type", match fn_decl.output {
+                FunctionRetTy::Default(_) => None,
+                FunctionRetTy::Ty(ty) => Some(ctx.intern(ty)),
+            })?;
+
+            let args: LuaResult<Vec<_>> = fn_decl.inputs
+                .into_iter()
+                .map(|arg| arg.into_lua_ast(ctx, lua_ctx))
+                .collect();
+
+            ast.set("args", iter_to_lua_array(args?.into_iter(), lua_ctx)?)?;
+
+            // TODO: self, self kind
+
+            Ok(ast)
+        })
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for Arg {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "Arg")?;
+        ast.set("id", self.id.as_u32())?;
+        ast.set("pat", self.pat.into_lua_ast(ctx, lua_ctx)?)?;
+
+        Ok(ast)
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for P<Block> {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "Block")?;
+
+        self.and_then(|block| {
+            let stmts = block.stmts
+                .into_iter()
+                .map(|stmt| stmt.into_lua_ast(ctx, lua_ctx))
+                .collect::<LuaResult<Vec<_>>>();
+
+            ast.set("stmts", iter_to_lua_array(stmts?.into_iter(), lua_ctx)?)?;
+
+            Ok(ast)
+        })
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for P<Pat> {
+    fn into_lua_ast(self, _ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "Pat")?;
+        ast.set("id", self.id.as_u32())?;
+
+        self.and_then(|pat| {
+            match pat.node {
+                PatKind::Wild => {
+                    ast.set("kind", "Wild")?;
+                },
+                PatKind::Ident(binding, ident, _sub_pattern) => {
+                    ast.set("kind", "Ident")?;
+                    ast.set("binding", match binding {
+                        BindingMode::ByRef(Immutable) => "ByRefImmutable",
+                        BindingMode::ByRef(Mutable) => "ByRefMutable",
+                        BindingMode::ByValue(Immutable) => "ByValueImmutable",
+                        BindingMode::ByValue(Mutable) => "ByValueMutable",
+                    })?;
+                    ast.set("ident", ident.as_str().get())?;
+                },
+                _ => return Err(LuaError::external(format!("unimplemented pattern type: {:?}", pat.node))),
+            }
+
+            Ok(ast)
+        })
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for Crate {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "Crate")?;
+        ast.set("module", self.module.into_lua_ast(ctx, lua_ctx)?)?;
+
+        Ok(ast)
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for Mod {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+        let items = self.items
+            .into_iter()
+            .map(|item| item.into_lua_ast(ctx, lua_ctx))
+            .collect::<LuaResult<Vec<_>>>();
+
+        ast.set("type", "Mod")?;
+        ast.set("inline", self.inline)?;
+        ast.set("items", iter_to_lua_array(items?.into_iter(), lua_ctx)?)?;
+
+        Ok(ast)
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for P<Item> {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "Item")?;
+
+        self.and_then(|item| {
+            ast.set("id", item.id.as_u32())?;
+            ast.set("ident", item.ident.name.as_str().get())?;
+
+            match item.node {
+                ItemKind::ExternCrate(opt_name) => {
+                    ast.set("kind", "ExternCrate")?;
+
+                    if let Some(name) = opt_name {
+                        ast.set("name", name.as_str().get())?;
+                    }
+                },
+                ItemKind::Mod(module) => {
+                    ast.set("kind", "Mod")?;
+                    ast.set("mod", module.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ItemKind::Use(_use_tree) => {
+                    ast.set("kind", "Use")?;
+                    // ast.set("use_tree", use_tree.into_lua_ast(ctx, lua_ctx))?;
+                },
+                ItemKind::Fn(decl, _header, _generics, block) => {
+                    ast.set("kind", "Fn")?;
+                    ast.set("decl", decl.into_lua_ast(ctx, lua_ctx)?)?;
+                    ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+                },
+                ItemKind::Struct(_variant_data, _generics) => {
+                    ast.set("kind", "Struct")?;
+                    // TODO: Variant data, generics
+                },
+                ItemKind::Impl(.., items) => {
+                    let items = items
+                        .into_iter()
+                        .map(|item| item.into_lua_ast(ctx, lua_ctx))
+                        .collect::<LuaResult<Vec<_>>>();
+
+                    ast.set("kind", "Impl")?;
+                    ast.set("items", iter_to_lua_array(items?.into_iter(), lua_ctx)?)?;
+
+                    // TODO: other fields
+                },
+                ref e => warn!("Item IntoLuaAst kind: {:?}", e),
+            }
+
+            Ok(ast)
+        })
+
+    }
+}
+
+impl<'lua> IntoLuaAst<'lua> for ImplItem {
+    fn into_lua_ast(self, ctx: &TransformCtxt, lua_ctx: LuaContext<'lua>) -> LuaResult<LuaTable<'lua>> {
+        let ast = lua_ctx.create_table()?;
+
+        ast.set("type", "ImplItem")?;
+
+        match self.node {
+            ImplItemKind::Method(sig, block) => {
+                ast.set("kind", "ImplMethod")?;
+                ast.set("ident", self.ident.to_string())?;
+                ast.set("decl", sig.decl.into_lua_ast(ctx, lua_ctx)?)?;
+                ast.set("block", block.into_lua_ast(ctx, lua_ctx)?)?;
+                // TODO: generics, attrs, ..
+            },
+            ref e => unimplemented!("IntoLuaAst for {:?}", e),
+        }
+
+        Ok(ast)
+    }
+}
